Fixed async/await, actually working
