+++ package.json
@@ -41,7 +41,7 @@
     "babel": "^5.8.29",
     "babel-eslint": "^4.1.1",
     "babel-runtime": "^5.8.29",
-    "eslint": "^1.8.0",
+    "eslint": "^1.9.0",
     "eslint-config-airbnb": "0.1.0",
     "npm-run-all": "^1.2.12",
     "rimraf": "^2.4.3"
