+++ test/fixture/config.json
@@ -1,0 +1,14 @@
+{
+  "podcasts": [
+    {
+      "name": "Reactive",
+      "podcastId": "1020286000",
+      "countries": ["us", "de", "gb", "se", "ca", "at", "au", "se", "nl", "br", "mx", "ru", "gr", "ar", "za", "ch", "pt"]
+    },
+    {
+      "name": "Descriptive",
+      "podcastId": "926224392",
+      "countries": ["us", "de", "gb", "se", "ca", "at", "au", "se", "nl", "br", "mx", "ru", "gr", "ar", "za", "ch", "pt"]
+    }
+  ]
+}
